Require python >= 3.9 to make it installable via poetry
Fixes the following two poetry-install errors:

```
The current project's supported Python range (>=2.7,<2.8 || >=3.4) is not compatible with some of the required packages Python requirement:
  - requests requires Python >=3.8, so it will not be installable for Python >=2.7,<2.8 || >=3.4,<3.8

Because surrealdb depends on requests (2.32.3) which requires Python >=3.8, version solving failed.
```

```
The current project's supported Python range (>=3.8) is not compatible with some of the required packages Python requirement:
  - aiohttp requires Python >=3.9, so it will not be installable for Python >=3.8,<3.9

Because surrealdb depends on aiohttp (3.11.11) which requires Python >=3.9, version solving failed.
```